+++ attributes/default.rb
@@ -19,10 +19,8 @@
 
 include_attribute "tomcat::default"
 
-version = '5.7.1'
-home_dir = "/usr/local/bonita-#{version}"
+default['bonita']['home_dir'] = "/usr/local/bonita"
 override['bonita']['version'] = version
-override['bonita']['home_dir'] = home_dir
 default['bonita']['license']['url'] = nil
 default['bonita']['license']['type'] = 'development'
 default['bonita']['license']['request'] = nil