+++ attributes/default.rb
@@ -1,9 +1,4 @@
 #
-# Cookbook Name:: bonita
-# Attributes:: default
-#
-# Copyright 2011, Peter Donald
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -17,31 +12,34 @@
 # limitations under the License.
 #
 
-include_attribute "tomcat::default"
+default['bonita']['home_dir'] = '/usr/local/bonita'
 
-default['bonita']['home_dir'] = "/usr/local/bonita"
+default['bonita']['user'] = nil
+default['bonita']['group'] = nil
+
 default['bonita']['license']['url'] = nil
 default['bonita']['license']['type'] = 'development'
 default['bonita']['license']['request'] = nil
 
-default['bonita']['packages']['bonita'] = nil
-default['bonita']['packages']['client'] = nil
-default['bonita']['packages']['keygen'] = nil
-default['bonita']['packages']['xcmis'] = nil
+default['bonita']['packages']['bonita_url'] = nil
+default['bonita']['packages']['client_url'] = nil
+default['bonita']['packages']['keygen_url'] = nil
 
-default['bonita']['extra_libraries'] = []
-default['bonita']['logging_properties'] = Mash.new
 default['bonita']['database']['hibernate']['dialect'] = 'org.hibernate.dialect.SQLServerDialect'
-default['bonita']['database']['exo_jcr']['dialect'] = 'mssql'
 default['bonita']['database']['hibernate']['interceptor'] = 'org.ow2.bonita.env.interceptor.MSSQLServerDescNullsFirstInterceptor'
 default['bonita']['database']['jdbc']['driver'] = 'net.sourceforge.jtds.jdbc.Driver'
+default['bonita']['database']['jdbc']['history']['resource'] = nil
 default['bonita']['database']['jdbc']['history']['url'] = nil
+default['bonita']['database']['jdbc']['history']['search_index_dir'] = '/usr/local/bonita/bonita/server/default/work/indexes/history'
 default['bonita']['database']['jdbc']['history']['schema'] = nil
+default['bonita']['database']['jdbc']['journal']['resource'] = nil
 default['bonita']['database']['jdbc']['journal']['url'] = nil
+default['bonita']['database']['jdbc']['journal']['search_index_dir'] = '/usr/local/bonita/bonita/server/default/work/indexes/journal'
 default['bonita']['database']['jdbc']['journal']['schema'] = nil
-default['bonita']['database']['jdbc']['xcmis']['url'] = nil
 default['bonita']['database']['jdbc']['username'] = nil
 default['bonita']['database']['jdbc']['password'] = nil
 
-default['bonita']['xcmis']['username'] = nil
-default['bonita']['xcmis']['password'] = nil
+default['bonita']['large_data_repository'] = '/usr/local/bonita/bonita/server/default/work'
+
+default['bonita']['tomcat']['extra_libraries'] = []
+default['bonita']['tomcat']['logging_properties'] = Mash.new