+++ metadata.rb
@@ -4,6 +4,6 @@ maintainer_email "peter@realityforge.org"
 license          "Apache 2.0"
 description      "Installs/Configures the bonita BPM server"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "5.7.1"
+version          "5.9.1"
 
-depends "tomcat"
+suggests "tomcat"